requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Pillow==8.3.1
pytesseract==0.3.8
pdf2image==1.16.0
nltk==3.6.2
torch==1.9.0
torchvision==0.10.0
transformers==4.9.2
Werkzeug==2.0.1